[BE] feat: 유동인구 pageable 기능 추가

+++ BackEnd/SpringBootServer/src/main/java/com/ssafy/backend/domain/district/entity/StoreDistrict.java
@@ -65,7 +65,7 @@ public class StoreDistrict {
 
     @Comment("개업률")
     @Column(columnDefinition = "Float")
-    private Long openedRate;
+    private Float openedRate;
 
     @Comment("폐업률")
     @Column(columnDefinition = "Float")
